# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cafe Calories API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/drivers/login": {
      "post": {
        "tags": [
          "drivers"
        ],
        "summary": "Driver Login",
        "operationId": "driver_login_drivers_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DriverLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/drivers/register": {
      "post": {
        "tags": [
          "drivers"
        ],
        "summary": "Driver Register",
        "operationId": "driver_register_drivers_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/drivers/me": {
      "get": {
        "tags": [
          "drivers"
        ],
        "summary": "Get Current Driver",
        "description": "Get the current authenticated driver's information.",
        "operationId": "get_current_driver_drivers_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/drivers/{driver_id}/assigned-orders": {
      "get": {
        "tags": [
          "drivers"
        ],
        "summary": "Get Assigned Orders",
        "operationId": "get_assigned_orders_drivers__driver_id__assigned_orders_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "driver_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Driver Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AssignedOrderOut"
                  },
                  "title": "Response Get Assigned Orders Drivers  Driver Id  Assigned Orders Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/drivers/{driver_id}/location": {
      "post": {
        "tags": [
          "drivers"
        ],
        "summary": "Post Location",
        "operationId": "post_location_drivers__driver_id__location_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "driver_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Driver Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DriverLocationIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/drivers/{driver_id}/location-status": {
      "post": {
        "tags": [
          "drivers"
        ],
        "summary": "Post Location With Status",
        "description": "Update driver location and status in one call.",
        "operationId": "post_location_with_status_drivers__driver_id__location_status_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "driver_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Driver Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DriverLocationWithStatus"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/drivers/{driver_id}/status": {
      "put": {
        "tags": [
          "drivers"
        ],
        "summary": "Update Driver Status",
        "description": "Update driver status (IDLE or OCCUPIED).",
        "operationId": "update_driver_status_drivers__driver_id__status_put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "driver_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Driver Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DriverStatusUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/drivers/available": {
      "get": {
        "tags": [
          "drivers"
        ],
        "summary": "Get Available Drivers",
        "description": "Get all idle drivers available for assignment.",
        "operationId": "get_available_drivers_drivers_available_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/IdleDriverInfo"
                  },
                  "type": "array",
                  "title": "Response Get Available Drivers Drivers Available Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/drivers/{driver_id}/orders/{order_id}/pickup": {
      "post": {
        "tags": [
          "drivers"
        ],
        "summary": "Pickup Order",
        "operationId": "pickup_order_drivers__driver_id__orders__order_id__pickup_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "driver_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Driver Id"
            }
          },
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignedOrderOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/drivers/{driver_id}/orders/{order_id}/status": {
      "post": {
        "tags": [
          "drivers"
        ],
        "summary": "Driver Update Order Status",
        "description": "Allow a driver to update the status of an order assigned to them.\nThis endpoint is driver-scoped and will validate assignment and valid transitions.",
        "operationId": "driver_update_order_status_drivers__driver_id__orders__order_id__status_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "driver_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Driver Id"
            }
          },
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Order Id"
            }
          },
          {
            "name": "new_status",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/OrderStatus"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignedOrderOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssignedOrderOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "cafe_id": {
            "type": "integer",
            "title": "Cafe Id"
          },
          "status": {
            "$ref": "#/components/schemas/OrderStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "total_price": {
            "type": "number",
            "title": "Total Price"
          },
          "total_calories": {
            "type": "integer",
            "title": "Total Calories"
          },
          "can_cancel_until": {
            "type": "string",
            "format": "date-time",
            "title": "Can Cancel Until"
          },
          "driver_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Driver Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "cafe_id",
          "status",
          "created_at",
          "total_price",
          "total_calories",
          "can_cancel_until",
          "driver_id"
        ],
        "title": "AssignedOrderOut"
      },
...
      "DriverLocationIn": {
        "properties": {
          "lat": {
            "type": "number",
            "title": "Lat"
          },
          "lng": {
            "type": "number",
            "title": "Lng"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "lat",
          "lng",
          "timestamp"
        ],
        "title": "DriverLocationIn"
      },
      "DriverLocationWithStatus": {
        "properties": {
          "lat": {
            "type": "number",
            "title": "Lat"
          },
          "lng": {
            "type": "number",
            "title": "Lng"
          },
          "status": {
            "$ref": "#/components/schemas/DriverStatus"
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "lat",
          "lng",
          "status"
        ],
        "title": "DriverLocationWithStatus"
      },
      "DriverLoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "DriverLoginRequest"
      },
      "DriverStatus": {
        "type": "string",
        "enum": [
          "IDLE",
          "OCCUPIED"
        ],
        "title": "DriverStatus"
      },
      "DriverStatusUpdate": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/DriverStatus"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "DriverStatusUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IdleDriverInfo": {
        "properties": {
          "driver_id": {
            "type": "integer",
            "title": "Driver Id"
          },
          "driver_name": {
            "type": "string",
            "title": "Driver Name"
          },
          "driver_email": {
            "type": "string",
            "title": "Driver Email"
          },
          "lat": {
            "type": "number",
            "title": "Lat"
          },
          "lng": {
            "type": "number",
            "title": "Lng"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "last_update": {
            "type": "string",
            "format": "date-time",
            "title": "Last Update"
          }
        },
        "type": "object",
        "required": [
          "driver_id",
          "driver_name",
          "driver_email",
          "lat",
          "lng",
          "status",
          "last_update"
        ],
        "title": "IdleDriverInfo"
      },
...
      "OrderStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "ACCEPTED",
          "DECLINED",
          "READY",
          "PICKED_UP",
          "CANCELLED",
          "REFUNDED",
          "DELIVERED"
        ],
        "title": "OrderStatus"
      },
...
      "Role": {
        "type": "string",
        "enum": [
          "USER",
          "OWNER",
          "STAFF",
          "ADMIN",
          "DRIVER"
        ],
        "title": "Role"
      },
      "Token": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "refresh_token"
        ],
        "title": "Token"
      },
      "UserCreate": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "password": {
            "type": "string",
            "title": "Password"
          },
          "role": {
            "type": "string",
            "title": "Role",
            "default": "User"
          },
          "height_cm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height Cm"
          },
          "weight_kg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weight Kg"
          },
          "dob": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dob"
          },
          "gender": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gender"
          },
          "activity_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Activity Level"
          },
          "daily_calorie_goal": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Daily Calorie Goal"
          }
        },
        "type": "object",
        "required": [
          "email",
          "name",
          "password"
        ],
        "title": "UserCreate"
      },
      "UserOut": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "required": [
          "email",
          "name",
          "id",
          "role",
          "is_active"
        ],
        "title": "UserOut"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/login"
          }
        }
      }
    }
  }
}